backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contractions==0.1.73
cryptography==45.0.6
cymem==2.0.11
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
huggingface-hub==0.34.3
idna==3.10
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.3
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyahocorasick==2.2.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.14.0
PyMuPDF==1.26.3
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
safetensors==0.6.1
semantic-version==2.10.0
setuptools-rust==1.11.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
sympy==1.13.3
tenacity==9.1.2
textsearch==0.0.24
thinc==8.3.6
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
